Split out time zone text from DateTimeTextProvider
