+++ core/src/commonMain/kotlin/io/islandtime/IslandTime.kt
@@ -2,6 +2,8 @@ package io.islandtime
 
 import io.islandtime.format.DateTimeTextProvider
 import io.islandtime.format.PlatformDateTimeTextProvider
+import io.islandtime.format.PlatformTimeZoneTextProvider
+import io.islandtime.format.TimeZoneTextProvider
 import io.islandtime.zone.PlatformTimeZoneRulesProvider
 import io.islandtime.zone.TimeZoneRulesProvider
 import kotlinx.atomicfu.atomic
@@ -15,6 +17,9 @@ object IslandTime {
 
     internal val dateTimeTextProvider: DateTimeTextProvider
         get() = settings.dateTimeTextProvider
+
+    internal val timeZoneTextProvider: TimeZoneTextProvider
+        get() = settings.timeZoneTextProvider
 
     @Suppress("ObjectPropertyName")
     private val _settings = atomic<Settings?>(null)
@@ -81,19 +86,26 @@ object IslandTime {
          * The date-time text provider to use.
          */
         var dateTimeTextProvider: DateTimeTextProvider
+
+        /**
+         * The time zone text provider to use.
+         */
+        var timeZoneTextProvider: TimeZoneTextProvider
     }
 
     private class InitializerImpl : Initializer {
         override var timeZoneRulesProvider: TimeZoneRulesProvider = PlatformTimeZoneRulesProvider
         override var dateTimeTextProvider: DateTimeTextProvider = PlatformDateTimeTextProvider
+        override var timeZoneTextProvider: TimeZoneTextProvider = PlatformTimeZoneTextProvider
 
         fun build(): Settings {
-            return Settings(timeZoneRulesProvider, dateTimeTextProvider)
+            return Settings(timeZoneRulesProvider, dateTimeTextProvider, timeZoneTextProvider)
         }
     }
 
     private data class Settings(
         val timeZoneRulesProvider: TimeZoneRulesProvider = PlatformTimeZoneRulesProvider,
-        val dateTimeTextProvider: DateTimeTextProvider = PlatformDateTimeTextProvider
+        val dateTimeTextProvider: DateTimeTextProvider = PlatformDateTimeTextProvider,
+        val timeZoneTextProvider: TimeZoneTextProvider = PlatformTimeZoneTextProvider
     )
 }
