+++ core/src/commonTest/kotlin/io/islandtime/format/DateTimeTextProviderTest.kt
@@ -1,7 +1,6 @@
 package io.islandtime.format
 
 import io.islandtime.DateTimeException
-import io.islandtime.TimeZone
 import io.islandtime.base.DateTimeField
 import io.islandtime.locale.localeOf
 import io.islandtime.test.AbstractIslandTimeTest
@@ -410,51 +409,5 @@ class DateTimeTextProviderTest : AbstractIslandTimeTest() {
             expected,
             DateTimeTextProvider.parsableTextFor(DateTimeField.ERA, TextStyle.FULL, en_US)
         )
-    }
-
-    @Test
-    fun `timeZoneTextFor() returns null when given a fixed offset time zone`() {
-        listOf(
-            TimeZone.FixedOffset("-04:00"),
-            TimeZone.FixedOffset("+00:00"),
-            TimeZone.FixedOffset("+14:00")
-        ).forEach { zone ->
-            TimeZoneTextStyle.values().forEach { style ->
-                assertNull(DateTimeTextProvider.timeZoneTextFor(zone, style, en_US))
-                assertNull(DateTimeTextProvider.timeZoneTextFor(zone, style, de_DE))
-            }
-        }
-    }
-
-    @Test
-    fun `timeZoneTextFor() returns a localized string when available`() {
-        val zone = TimeZone("America/New_York")
-
-        assertEquals(
-            "Eastern Standard Time",
-            DateTimeTextProvider.timeZoneTextFor(zone, TimeZoneTextStyle.STANDARD, en_US)
-        )
-        assertEquals(
-            "EST",
-            DateTimeTextProvider.timeZoneTextFor(zone, TimeZoneTextStyle.SHORT_STANDARD, en_US)
-        )
-        assertEquals(
-            "Eastern Daylight Time",
-            DateTimeTextProvider.timeZoneTextFor(zone, TimeZoneTextStyle.DAYLIGHT_SAVING, en_US)
-        )
-        assertEquals(
-            "EDT",
-            DateTimeTextProvider.timeZoneTextFor(zone, TimeZoneTextStyle.SHORT_DAYLIGHT_SAVING, en_US)
-        )
-    }
-
-    @Test
-    fun `timeZoneTextFor() returns null when the zone is invalid`() {
-        val zone = TimeZone("America/Boston")
-        assertFalse { zone.isValid }
-
-        TimeZoneTextStyle.values().forEach { style ->
-            assertNull(DateTimeTextProvider.timeZoneTextFor(zone, style, en_US))
-        }
     }
 }
