+++ core/src/jvmTest/kotlin/io/islandtime/format/JvmTimeZoneTextProviderTest.kt
@@ -7,7 +7,7 @@ import org.junit.Test
 import kotlin.test.assertNull
 
 @Suppress("PrivatePropertyName")
-class JvmDateTimeTextProviderTest : AbstractIslandTimeTest() {
+class JvmTimeZoneTextProviderTest : AbstractIslandTimeTest() {
     private val en_US = localeOf("en-US")
     private val de_DE = localeOf("de-DE")
 
@@ -15,7 +15,7 @@ class JvmDateTimeTextProviderTest : AbstractIslandTimeTest() {
     fun `timeZoneTextFor() returns null for generic styles - limited by Android currently`() {
         val zone = TimeZone("America/New_York")
 
-        assertNull(DateTimeTextProvider.timeZoneTextFor(zone, TimeZoneTextStyle.GENERIC, en_US))
-        assertNull(DateTimeTextProvider.timeZoneTextFor(zone, TimeZoneTextStyle.SHORT_GENERIC, de_DE))
+        assertNull(TimeZoneTextProvider.timeZoneTextFor(zone, TimeZoneTextStyle.GENERIC, en_US))
+        assertNull(TimeZoneTextProvider.timeZoneTextFor(zone, TimeZoneTextStyle.SHORT_GENERIC, de_DE))
     }
 }
